requirements.txt
# MVP Auto-Summary Orchestrator
# Replaces n8n with native Python scheduler + tasks

# Database
psycopg2-binary==2.9.9

# HTTP client (async-capable, connection pooling)
httpx>=0.27.0,<1.0

# Scheduler (cron-like job scheduling)
APScheduler>=3.10.4,<4.0

# Telegram bot framework (async, long-polling)
python-telegram-bot>=21.0,<22.0

# Configuration from .env
pydantic-settings>=2.0,<3.0
python-dotenv>=1.0.0,<2.0

# Structured logging
structlog>=24.0.0,<25.0

# Retry with exponential backoff
tenacity>=8.2.0,<10.0

# Testing
pytest>=8.0.0,<9.0
pytest-asyncio>=0.23.0,<1.0
